prepare admin: magang kegiatan, refactor mahasiswa
