+++ routes/web.php
@@ -1,11 +1,13 @@
 <?php
 
+use App\Http\Controllers\AdminMagangController;
 use App\Http\Controllers\AdminProfilAdminController;
 use App\Http\Controllers\AdminProfilDosenController;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\DosenController;
 use App\Http\Controllers\MahasiswaAkunProfilController;
 use App\Http\Controllers\MahasiswaMagangController;
+use App\Http\Controllers\MahasiswaPengajuanController;
 use App\Http\Controllers\ProgramStudiController;
 use App\Http\Controllers\PerusahaanMitraController;
 use Illuminate\Support\Facades\Auth;
@@ -73,6 +75,9 @@ Route::middleware(['auth'])->group(function () {
         Route::put('/admin/perusahaan/{id}', [PerusahaanMitraController::class, 'update']);
         Route::delete('/admin/perusahaan/{id}', [PerusahaanMitraController::class, 'destroy']);
         Route::patch('/admin/perusahaan/{id}/toggle-status', [PerusahaanMitraController::class, 'toggleStatus'])->name('admin.toggle-status');
+
+        // MAGANG: Keagiatan
+        Route::get('/admin/magang/kegiatan', [AdminMagangController::class, 'kegiatan']);
     });
 
     Route::middleware(['authorize:dosen'])->group(function () {
@@ -85,31 +90,24 @@ Route::middleware(['auth'])->group(function () {
 
     Route::middleware(['authorize:mahasiswa'])->group(function () {
         Route::get('/mahasiswa', [MahasiswaAkunProfilController::class, 'index'])->name('mahasiswa.index');
+        // PROFILE
         Route::get('/mahasiswa/profile', [MahasiswaAkunProfilController::class, 'profile'])->name('mahasiswa.profile');
         Route::get('/mahasiswa/profile/edit', [MahasiswaAkunProfilController::class, 'profile'])->name('mahasiswa.profile.edit');
         Route::post('/mahasiswa/profile/update', [MahasiswaAkunProfilController::class, 'update'])->name('mahasiswa.profile.update');
         Route::post('/mahasiswa/profile/update-password', [MahasiswaAkunProfilController::class, 'changePassword'])->name('mahasiswa.profile.update-password');
         Route::get('/mahasiswa/dokumen', [MahasiswaAkunProfilController::class, 'dokumen'])->name('mahasiswa.dokumen');
         Route::post('/mahasiswa/dokumen/upload', [MahasiswaAkunProfilController::class, 'dokumenUpload'])->name('mahasiswa.dokumen.upload');
-        Route::get('/mahasiswa/magang', [MahasiswaMagangController::class, 'magang'])->name('mahasiswa.magang');
+        // MAGANG
+        Route::get('/mahasiswa/magang', [MahasiswaMagangController::class, 'index'])->name('mahasiswa.magang');
         Route::get('/mahasiswa/magang/lowongan/', function () {
             return redirect('/mahasiswa/magang');
         });
         Route::get('/mahasiswa/magang/lowongan/{lowongan_id}', [MahasiswaMagangController::class, 'magangDetail'])->name('mahasiswa.magang.lowongan.detail');
         Route::get('/mahasiswa/magang/lowongan/{lowongan_id}/ajukan', [MahasiswaMagangController::class, 'ajukan'])->name('mahasiswa.magang.lowongan.ajukan');
         Route::post('/mahasiswa/magang/lowongan/{lowongan_id}/ajukan', [MahasiswaMagangController::class, 'ajukanPost'])->name('mahasiswa.magang.lowongan.ajukan.post');
-        Route::get('/mahasiswa/magang/pengajuan', [MahasiswaMagangController::class, 'pengajuan'])->name('mahasiswa.magang.pengajuan');
-        Route::get('/mahasiswa/magang/pengajuan/{pengajuan_id}', [MahasiswaMagangController::class, 'pengajuanDetail'])->name('mahasiswa.magang.pengajuan.detail');
-        Route::delete('/mahasiswa/magang/pengajuan/{pengajuan_id}', [MahasiswaMagangController::class, 'pengajuanDelete'])->name('mahasiswa.magang.pengajuan.delete');
-    });
-
-    Route::prefix('admin/perusahaan')->group(function () {
-        Route::get('/', [PerusahaanMitraController::class, 'index'])->name('perusahaan.index');
-        Route::post('/list', [PerusahaanMitraController::class, 'list'])->name('perusahaan.list'); // untuk DataTables AJAX
-        Route::get('/create', [PerusahaanMitraController::class, 'create'])->name('perusahaan.create');
-        Route::post('/store', [PerusahaanMitraController::class, 'store'])->name('perusahaan.store');
-        Route::get('/{id}/edit', [PerusahaanMitraController::class, 'edit'])->name('perusahaan.edit');
-        Route::put('/{id}/update', [PerusahaanMitraController::class, 'update'])->name('perusahaan.update');
-        Route::delete('/{id}/delete', [PerusahaanMitraController::class, 'destroy'])->name('perusahaan.destroy');
+        // PENGAJUAN
+        Route::get('/mahasiswa/magang/pengajuan', [MahasiswaPengajuanController::class, 'index'])->name('mahasiswa.magang.pengajuan');
+        Route::get('/mahasiswa/magang/pengajuan/{pengajuan_id}', [MahasiswaPengajuanController::class, 'pengajuanDetail'])->name('mahasiswa.magang.pengajuan.detail');
+        Route::delete('/mahasiswa/magang/pengajuan/{pengajuan_id}', [MahasiswaPengajuanController::class, 'pengajuanDelete'])->name('mahasiswa.magang.pengajuan.delete');
     });
 });
